+++ misi_hijau/res/res.pyxres
